requirements.txt
numpy~=1.21.5
ujson == 1.35
paramiko == 2.8.0
pycryptodome == 3.11.0
opencv-python-headless >= 4.5.4.60
scp == 0.14.2